Allow services assigned via hostgroup_name to be explicitly overridden
With this patch, a service assigned to a host via one of that host's
hostgroup memberships is considered a low precedence assignment. If
another service with the same service_description is assigned to the
same host using the host's host_name for assignment, that service
will be used instead.

This is a handy way of modifying a few parameters for a service-check
on a host that is otherwise monitored exactly like the rest of the
hosts in its hostgroup without having to bother with explicitly
excluding the specially treated host(s) in the service object
definition.

Signed-off-by: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1264 5f96b256-904b-4d8d-8c98-d829582c6739