base/commands: Fix error path of opening checkresult files
Previously we used to check that the returned filedescriptor
was larger than 0 to consider it a success and silently return
if it wasn't, but open() (as all other system calls) can return
0 on a successful invocation. Previously, this resulted in one
lost filedescriptor and the memory used for storing the path
to the checkresult file being leaked.

This patch takes care of both those leaks.

Signed-off-by: Andreas Ericsson <ae@op5.se>


Author: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1501 5f96b256-904b-4d8d-8c98-d829582c6739