threadsafaty: Use reentrant versions of localtime() and gmtime()
This makes Nagios play well with Merlin if macros are exported
to environment variables.

If/when someone complains about Nagios not compiling on their
system, we'll introduce lock-wrapped versions of these functions
for use on those systems, but it's unlikely that a system that
supports threads doesn't have gmtime_r() and localtime_r() readily
available.

Reported-by: Patrik Båt <patrik.bat@cypoint.se>
Signed-off-by: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1533 5f96b256-904b-4d8d-8c98-d829582c6739