Optionally allow assigning services to empty hostgroups
We used to bomb out with an error on such configurations, but with
this option it becomes optional if we should produce an error or a
warning.

It's useful to be able to assign services to empty hostgroups when
configuration files or pre-cached object files are distributed to
various pollers, or when the process of generating Nagios config is
automated, or when a set of services is slowly being phased out but
should be kept around in case of 'oh shit' epiphanies, such as one
might experience when phasing out old monitoring agents.

Thanks to Comcast in general and Max Schubert in particular for
sharing this patch.

Authored-by: Max Schubert <maxs@webwizarddesign.com>
Reviewed-by: Daniel Wittenberg <daniel.wittenberg.r0ko@statefarm.com>
Reviewed-by: Michael Friedrich <michael.friedrich@univie.ac.at>
Signed-off-by: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1554 5f96b256-904b-4d8d-8c98-d829582c6739