Fix form/table DOM confusion in cgi's
The CGI's behave well enough in most cases, but when loaded inside
another layer (such as by a <div> tag), some of the forms become
inaccessible due to invalid html.

Typicall, forms are started inside tables but outside a vertical
table delimiter, and then ended outside the table altogether.

This causes confusion for some browsers, so let's avoid that
sort of thing and start and end the forms outside the tables
or, where appropriate, inside the same table delimiter as the
form started.

Signed-off-by: Sven-Göran Bergh <sgb@systemasis.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>


Author: Sven-Göran Bergh <sgb@systemasis.com>
Committer: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1661 5f96b256-904b-4d8d-8c98-d829582c6739