trends-add-forward-to-next-problem.patch
In order to allow for scanning through the remaining recording and
searching for the occurrence of the next host or service problem,
the addtional time period tag TIMEPERIOD_NEXTPROBLEM was introduced.
If selected, the time period of the next diplay will be selected in
such a way that the full duration of the next problem will be
displayed at the center of the image. If only the start time is
available, because the problem is still present, the time window
of the image falls back to the 24-hour period.

A small fix was required to solve an inconsistency of the image
width (600 vs. 900 pixels).

Signed-off-by: Carsten Emde <cbe@osadl.org>


Author: Carsten Emde <C.Emde@osadl.org>
Committer: Andreas Ericsson <ae@op5.se>


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1662 5f96b256-904b-4d8d-8c98-d829582c6739