Handle timeperiods spanning across dst changes gracefully
Previously Nagios used to balk on "special" timeperiods during
changes in daylight savings time. This patch adds a check to
see if we're crossing the DST line and, depending on "direction"
adds or subtracts 3600 seconds to the calculated time.

This patch reportedly works well for Wuerth-Phoenix, according
to Stefano Di Luca.

Fixed-by: Stefano Di Luca <Luca.DiStefano@wurth-phoenix.com>
Signed-off-by: Stefano Di Luca <Luca.DiStefano@wurth-phoenix.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1726 5f96b256-904b-4d8d-8c98-d829582c6739