core: Make command_check_interval=-1 equal command_check_interval=5s
The docs claim that "-1" is supposed to mean "check as often as
possible", but the code begs to differ and uses a command check
interval of 60 seconds in such cases. I have no idea why, but it
appears the super-unlikely has happened and the docs got updated
but not the code instead of the other way around.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1742 5f96b256-904b-4d8d-8c98-d829582c6739