core: Avoid using global macros when sending notifications
It makes no sense to have a single spot apart from initialization
where the global macro struct is used. It doesn't even make much
sense there, but at least we can get rid of as much as possible
of the global madness.

This might fix the problem of unassigned contacts receiving
notifications, but requires rigorous testing.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1744 5f96b256-904b-4d8d-8c98-d829582c6739