core: Avoid sending notifications to the wrong contact
When a nebmodule blocks a notification, notification_list wasn't
cleared, causing the next notification to go out to all contacts for the
blocked notification, as well as to all contacts for the current
notification.

Fix by remember to clear notification_list after a nebmodule returns,
and by always clearing notification_list prior to filling it up again to
avoid similar bugs in the future. It's not dangerous to call too often,
and it's very cheap when notification_list is NULL.

Patch-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1745 5f96b256-904b-4d8d-8c98-d829582c6739