core: Plug memory leak on display_name variables
Previously we used to leak the display_name once per config reload.
For small configurations or, obviously, where display_name isn't
used, this is not a problem. On larger configurations with periodic
frequent reloads this could be problematic.

We also used to leak the display_name once on each error, which we
no longer do although that is basically a non-issue since errors
always cause us to bomb out early anyway.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1794 5f96b256-904b-4d8d-8c98-d829582c6739