core: fix possible race condition on downtime list lock
When deleting a downtime, the newly introduced nagios_downtime_lock
mutex gets locked to prevent threads from fiddling in there.

After having fetched the downtime we want to delete we only try to
if !NULL and then releasing the mutex.

If a module has already manipulated the list in memory and deleted
the downtime, the condition runs into the else-tree returning an
error and the mutex will be kept locked forever.

In order to fix that we move the unlocking outside the if condition,
let memory being free'd or result being an error and then savely
unlock the downtime list.

This will prevent us from running into a possible race condition
leaving the core idle wait for the mutex on deleting a comment.

Pached-by: Michael Friedrich <michael.friedrich@univie.ac.at>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1839 5f96b256-904b-4d8d-8c98-d829582c6739