core: Avoid insane looping when rescheduling checks
Previously we used to iterate over the entire event queue to
see if the object we were about to reschedule a check for was
already scheduled or not. Since checks are normally rescheduled
just after they've been run, that means we never hit an event
inside the list and so we walk all of it without ever finding
it, causing O(n) worst-case complexity and a probability of
hitting that worst-case complexity of very nearly 100%.

With this patch, we reuse the currently unused nexthash field
of hosts and service objects to stash the next scheduled check
event, providing constant-time access to it and avoiding the
insane looping, which should reduce Nagios' cpu consumption
substantially, since almost 80% of Nagios' user cpu time is
spent on this loop.

Relevant gprof output before patch:
% time  cumulat   self   calls  ms/call ms/call name
 17.24   0.05     0.05   3490   0.01     0.01   schedule_service_check
 17.24   0.10     0.05    1    50.00   261.97   event_execution_loop

Relevant gprof output after patch:
% time  cumulat   self   calls  ms/call ms/call name
 12.20   0.05     0.05       1  50.00  352.26   event_execution_loop
  9.76   0.09     0.04   10565   0.00    0.00   clear_host_macros_r
....
  0.00   0.41     0.00    3594   0.00    0.01   schedule_service_check

From 17.24% of the time used to less than 0.01 is quite good,
so we stick with that and let it be enough. Testing is ofcourse
required to make sure checks aren't being run extremely frequently,
but that shouldn't be the case since the new check is assigned to
the scheduled object in case we replace it with a forced one.

Reported-by: Mathias Kettner <mk@mathias-kettner.de>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1850 5f96b256-904b-4d8d-8c98-d829582c6739