core: Fix event removal from queues
The primary feature of doubly linked lists is insertion and
removal in O(1) time, but the remove_event() function used
an O(n) algorithm for removal, traversing the entire chain
to locate the event to remove when it was already given.

In the normal case, the event to remove is at the top of the
list, so the "walk it all to actually find something" only
triggers when the user schedules a forced check via commands,
but the fix is trival so we might as well apply it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1851 5f96b256-904b-4d8d-8c98-d829582c6739