macros: Avoid senseless looping when free()'ing macros
This patch replaces a bunch of for() loops with an inside switch()
clause filtering in only macros we know we should remove with an
unrolled loop removing only those macros we already know we want
to remove anyways.

It's possible but unlikely that a clever c-compiler already does
this optimization, but the current gprof output after 3490 service
checks indicates that at least gcc 4.5.1 doesn't do this.

Relevant part of gprof output:
time   seconds   seconds   calls  ms/call  ms/call  name
 12.20     0.05    0.05       1    50.00   352.26  event_execution_loop
  9.76     0.09    0.04    10565    0.00     0.00  clear_host_macros_r

Although clear_host_macros_r() is the largest consumer of cpu time
here (as it's a child of event_execution_loop, it's time will
propagate up to it), we do the same fix for all clear_*_macro
functions to avoid all unnecessary loops.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1853 5f96b256-904b-4d8d-8c98-d829582c6739