core: Revamp file_uses_embedded_perl()
Primaily we do this to get rid of the non-reentrant strtok() call.
Since we only handle one type of directive anyway, we can easily
optimize for that simple case and be done much quicker. If we add
more directives later we handle that then.

While we're at it, we get rid of some superfluous variables to
reduce the memory footprint of Nagios somewhat and generally clean
up the code.

We also remove the completely insane indentation depth that used
to be there that was exceedingly difficult to keep track of.

To top it all off, we increase the buffer we're reading in to. A
lot of the Perl plugins on Nagios-exchange use lines longer than
80 characters, so if we fgets() such a line the "nagios directives
must be within first 10 lines" is no longer true. None of the (few)
plugins I checked has any top-level comments spanning more than
256 chars though, and since that's a nice round number (for stuff
counting things in binary anyways) we stick with that.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1854 5f96b256-904b-4d8d-8c98-d829582c6739