core: Clear the escalation macro even if there are no contacts
This clears up a microscopic memory leak that would previously occur
when a service notification was triggered but no valid contacts were
found to send the notifications to.

It also happened for all host notifications, but since the logic is
the same, we simply copy the solution we use for services and be done
with it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1901 5f96b256-904b-4d8d-8c98-d829582c6739