core: Revamp how we handle raw command lines
Previously we used to free() the raw_command in a bunch of different
places to account for all the different things we tried to do, but
in a few places we could still end up forgetting to do so (notably
if we failed to produce a processed command line for checks).

Since the raw commandline is only ever used to produce the processed
command line, we can free() it immediately after we've done so and
ignore the results of process_command_line_r() to avoid having to
keep track of a variable that's never used again.

For consistency's sake, we make the same change everywhere where we
compute a command line.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1905 5f96b256-904b-4d8d-8c98-d829582c6739