core: Unify check scheduling replacement logic
Previously it was quite hard to follow and would sometimes allocate
memory even when it wouldn't be used. With this patch, we make the
allocation never happen unless it's needed, and we make sure all
objects are updated with the proper data.

This might fix the issue where service checks are being run both
at the override time (for forced checks) and the originally
scheduled time, although I haven't tested that.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1958 5f96b256-904b-4d8d-8c98-d829582c6739