macros: Try the most common macros first
The current way of checking if a macro is of a certain type is to do
a linear scan of all X-macros first, followed by ARG and, lastly,
the USER macros. That's pretty backwards, since ARG macros are by far
the most commonly used ones, outnumbering all others combined by a
factor of 3.4 (in a pretty average test-config). This leads to a lot
of unnecessary strcmp() calls which we certainly can do without.

With this patch, we go directly to the ARG macros, shortening the path
somewhat. We also avoid allocating any memory if the macro we're after
can be returned without extra allocs.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1959 5f96b256-904b-4d8d-8c98-d829582c6739