macros: Use binary search when looking up macro names
We used to use a linear search when we encountered a standard macro
(rather, when we encountered a non-USER and non-ARG macro) over the
list of macro names. For non-hits, that resulted in 152 strcmp()'s,
and an average count of 76 strcmp()'s.

With this patch, we take that down to a max of 7 strcmp()'s and some
fast integer math.

To be honest though, the most common "x" macro used is $HOSTADDRESS$,
which we've actually worsened the case for somewhat, so we add a
shortcut for that particular macro.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1960 5f96b256-904b-4d8d-8c98-d829582c6739