Revert "Janitor patch: Fix a bunch of -Wset-but-unused warnings"
This reverts commit 8e6295d295dfeec0f884e5e45a17cb22b05921cd,
which apparently broke macro resolution somehow.

This seems to indicate that gcc did something wrong somehow,
which is pretty worrying in itself, but since I'm currently lazy
and it's far easier to live with compiler warnings than to locate
the one triggering this bug, the extra 100 or so bytes of stack-
space these variables waste will have to be wasted for a while
longer.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1965 5f96b256-904b-4d8d-8c98-d829582c6739