core: Use priority queue based event queue instead of linked lists
The linked lists have O(n) insertion time, worst case. The theory
is that this worst case scenario shouldn't happen all too often,
since events should, most of the time, be scheduled after all
other events that are already scheduled. In a real-world scenario
with 2500 hosts and 55000 service checks, with all scheck scheduled
at a 5 minute interval, this failed pretty hard due to the linear
behaviour of the old algorithm.

With the new squeue-on-top-of-pqueue implementation, we're
guaranteed O(lg n) behaviour for insertion and removal.

Although pop() (extract next event) moves from O(1) to O(lg n)
due to the need for bubbling of remaining events, we still make
a *huge* win due to this change, and get a far more predictable
behaviour even when we run into larger configurations where
check scheduling may not be a fixed interval for all checks.

The relevant profiling data from two runs of ~5 minutes each
(discrepancies in runtime is due to Nagios shutdown time) is as
follows:

Before:
  %      self              self     total
 time    seconds   calls   s/call   s/call  name
 74.57    47.35    70563     0.00     0.00  add_event

After:
  %      self              self     total
 time    seconds   calls   s/call   s/call  name
  0.00     0.00    76782     0.00     0.00  add_event

From 74.57% of the allotted cpu-time over a 5-minute period to an
immeasurably small amount is pretty good, even if I do say so
myself ;)

For smaller configurations, the benefits will be negligent and
the performance gains become more noticeable the larger the setup.

A downside of this patch is that the sizes and layouts of the two
structs timed_event and scheduled_downtime have changed. Since
none of the four major eventbroker modules (NDOUtils, Livestatus,
mod_gearman and Merlin) use them or are affected in any way, I
really don't care about that at this point in time.

Thanks-to: Volkan Yazici <volkan.yazici@gmail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1971 5f96b256-904b-4d8d-8c98-d829582c6739