nagiostats: Make the -s flag useful once more
It used to be used for reading some format that no longer exists,
but the helptext indicated it was supposed to be used to read a
normal status.log file (and thus just bypass the nagios.cfg).
Since there are no tools to generate the now-dead format, we
simply re-use the -s flag to mean "status.log format".

As a nice side-effect, we can remove quite a lot of dead code.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1976 5f96b256-904b-4d8d-8c98-d829582c6739