core: Fix checking of return-value for schedule_new_event()
Since we changed it to return a pointer to a timed_event, we've been
getting compiler warnings from cmd_signal_process().

This patch silences the warning by properly checking the return value
of schedule_new_event() and returning properly based on its result.

We still ignore the result at the original callsite, but at least now
we get rid of the "pointer-to-int conversion" warning gcc used to
throw at us.

Patched-by: Ricardo Maraschini <ricardo.maraschini@opservices.com.br>
Message-by: Andreas Ericsson <ae@op5.se>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1978 5f96b256-904b-4d8d-8c98-d829582c6739