lib/test-squeue: Compute expected event priority after adding
The squeue library can (and does) sometimes change the time we pass
it as for when the event is supposed to run, which would cause a
failed testrun about once in 100 runs (probably due to the internal
clock switching second).

This patch fixes it, using the adjustments the squeue api does for
us to calculate the expected priority.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1997 5f96b256-904b-4d8d-8c98-d829582c6739