lib/iocache: Make iocache_use_delim() properly mark all data used
It makes no sense to force callers to fiddle with adding or
subtracting the delimiter, so we remove it from the size calculation
and make sure we just pass the delimiter length along when marking
data as used in the io cache.

Tests are naturally fixed to reflect the new world order.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2002 5f96b256-904b-4d8d-8c98-d829582c6739