lib/iocache: Fix iocache_resize()
Not using the new buffer and not updating the bufsize was pretty
damn stupid, to tell the truth. In all honesty, we should probably
return some iocache specific error in case someone's trying to shrink
the buffer to less than iocache_available(ioc), in order to prevent
data losses.

OTOH, that's really up to the caller.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2004 5f96b256-904b-4d8d-8c98-d829582c6739