lib/kvvec:buf2kvvec() Fix key/value pair counting (again)
The last patch assumed we would send the delimiter to the buf2kvvec()
machinery, and that the delimiter was the same as the pair separator.

Since that's an obvious bogosity in a standalone library, we fix that
here and will instead make certain we only send the buffer we want
buf2kvvec() to operate on.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2006 5f96b256-904b-4d8d-8c98-d829582c6739