libnagios: Add runcmd and worker libraries
Although the worker library isn't, strictly speaking, a library as
much as it is a toolkit and some application logic, it's nifty to
have for creating remote workers or as an example on how to use the
other parts of libnagios.

The runcmd library is inherited from the plugins project and might
need to be refactored for performance later.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2009 5f96b256-904b-4d8d-8c98-d829582c6739