core: Handle brokered I/O events adding high priority events
Previously we used to go ahead with the event we got from the
queue no matter if events brokered from the I/O broker turned
were added that were supposed to be run before that one. We still
do (sort of), except that now we prioritize the earlier scheduled
events if the first event is scheduled far enough into the future.

"Far enough" means "one second" in this case.

While we're at it, we remove the fairly useless wproc_poll() call.
Since nagios_iobs is a global variable and wproc_poll() is just a
wrapper around iobroker_poll() called with nagios_iobs and the
timeout as variables, we really don't need the wrapper.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2019 5f96b256-904b-4d8d-8c98-d829582c6739