core: Process check results ad-hoc
First relegating them to an in-memory list is just an unnecessary
extra step that we can avoid entirely by just passing the parsed
check results to their respective handlers as soon as we encounter
them. While we might encounter occasional long timeouts while reaping
checks (in case they cause notifications or other helper programs to
be spawned), we will save some cpu time overall on this change, and
some extra memory as well.

Also, we avoid the whole problem of potentially reaping more checks
than we have time to process later, which could cause a buildup of
check result data in memory which was possible (although very unlikely)
with the old code.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2036 5f96b256-904b-4d8d-8c98-d829582c6739