Retire periodic command checking
Since commands are always processed immediately upon arrival since
the inception of the command file worker thread, there's really no
need to maintain variables that never get filled.

This commit was intended to be squashed with the previous one first,
but since it's so large on its own I'll leave it as-is.

In essence, we remove the two global variables "command_check_interval"
and "last_command_check_time", and all code affected by that removal,
as well as example configurations and test-data.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2046 5f96b256-904b-4d8d-8c98-d829582c6739