core: Remove threading
Since the introduction of the command file worker no threads are used
and full parallalization is acquired by means of helper processes and
message passing instead, exactly like a micro-kernel would do it.

We take some small overhead in the message passing, but it's negligible
in the big picture and can, if necessary, be made faster through more
complicated mechanisms, such as binary protocols or shared memory.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2049 5f96b256-904b-4d8d-8c98-d829582c6739