include/nagios: Re-update path to libnagios.h
In order to be able to install the nagios headers as a development
package the lib path really must be relative to the -I flag we
provide when compiling. Since we always *do* provide that -I flag
but can't do that when other people are building their packages
against our devel packages (neb-authors), we'll have to install
the devel packages in "/usr/include/nagios/lib/" or some such, and
then a relative path will work.

Users who can't get the path to work right now will have to re-run
./configure (which is now also mentioned in a comment right next
to the #include so as to alert the unwary developer...).

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2054 5f96b256-904b-4d8d-8c98-d829582c6739