Combine common macro declarations
Both cgi's and core hade STATE_WARNING etc defined in their own
headers, but with identical values (and comments). This patch just
moves them to the common header and leaves it at that.

Disentangling the fact that HOST_UNREACHABLE etc has different
values is another beast best left to the glorious day when we no
longer need to worry about the UI at all.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2060 5f96b256-904b-4d8d-8c98-d829582c6739