Janitor patch: Clear up global and unused variable
A lot of global variables are declared in files that doesn't use
them, and a lot of variables were set but not used, leading to
wasted cycles when using less clever compilers than gcc 4.x.

This patch clears up the cruft, and manages to kill some of the
(very many) global variables.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2063 5f96b256-904b-4d8d-8c98-d829582c6739