lib/iocache: Add method for moving iocache offset pointer backwards
A common pattern in serialized data is to start a message with a header
that specify the message size. To make iocache more suitable for that
kind of use, we want to be able to first read the header, then see if
a body is available, and if things muck up, put the header back where we
found it. That means we have to be able to subtract the offset, which is
what this patch enables.

Patched-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2066 5f96b256-904b-4d8d-8c98-d829582c6739