Add 'install-devel' as a top-level make target
This patch provides Naigos with a devel package (which is intended
to grow over time). The idea is to let module and addon authors use
this in order to access Nagios' data rapidly with a minimum of code
duplication.

Modules should *not* link to libnagios however, as it's only
provided as a static library and is already linked into Nagios
itself.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2070 5f96b256-904b-4d8d-8c98-d829582c6739