Add lib/lnag-utils.h
Since libnagios can't use anything from core nagios although the
other way around works fine, we're forced to move everything we
want to have in the library to somewhere where we can use it. The
things we want to move includes a few macros that are generally
useful, so we introduce the "lnag-utils.h" header file, which can
keep track of declarations for small inline functions, various
micro-helpers and the above-mentioned useful macros for us, while
updating everything else to be able to make use of it.

This means the CGI's can now use header-files from libnagios exactly
like the core daemon can, which paves the way for linking the cgi's
against libnagios too.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2078 5f96b256-904b-4d8d-8c98-d829582c6739