libnagios: Include the skiplist API in libnagios and link cgi's to it
This means we no longer have to link the skiplist object file
explicitly with either nagios core or the cgi's, so we tuck in the
build infrastructure to make that happen. We also move a few more
Nagios-centrict macros to lnag-utils (TRUE, FALSE, ERROR and OK) to
prevent the skiplist library having to depend on files outside the
lib directory.

While we're at it, we improve a bit on the library and make two
private helpers static and removing their prototypes from the header
file.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2079 5f96b256-904b-4d8d-8c98-d829582c6739