Add bitmap library to libnagios
This is insanely useful for calculating dependency graphs now that
each object has a unique id, and will briefly be used to do exactly
that.

Actually, it's insanely useful to flag just about anything with an
id in constant time and very little space, but we'll initially be
using it to check for circular paths in service dependencies.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2081 5f96b256-904b-4d8d-8c98-d829582c6739