libnagios: Add dkhash, the dual-keyed hash api
Since most objects in Nagios are uniquely identified by two keys
(as opposed to just one), it's prudent to make it easy to give
quick access to such objects with simple API calls, both to Nagios
and to all its modules.

This library will be used as the basis for object lookups in the
near future, as it provides O(1) lookup time, to be compared with
the O(n lg n) time provided by skiplists.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2082 5f96b256-904b-4d8d-8c98-d829582c6739