Store backups when running indent
It's highly annoying to accidentally run the indentation when one's
in the middle of something and not all braces are closed. In theory,
it would be enough to just close the braces and undo the indent,
but in practice it's not quite so simple.

With this patch, astyle at least creates backup files suffixed with
.pre-indent, which lets one easily recover from such a mistake
without having to spend several hours to do it, like I just did
(and still do...).

Thank god for git which let me get the unmodified files back in
just about notime at least.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2094 5f96b256-904b-4d8d-8c98-d829582c6739