lib/lnag-utils: Fix ispof2() and rup2pof2()
They weren't cryptic enough before, but with an additional tertiary
statement we get them to be completely illegible for anyone without
an advanced degree in computer science. It also makes them correct,
since we previously claimed that 0 and 1 were prime, and tried to
round up 1 to 1 when rounding to the nearest power of two.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2097 5f96b256-904b-4d8d-8c98-d829582c6739