lib/dkhash: Major fixes
We now check for and disallow duplicates in all hash tables and fix
tests accordingly. The hash function itself has been tweaked to provide
better dispersion for a superset of the default checkcommand list.
More tuning might prove useful, but we can do that if we get some real
data to test it on.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2098 5f96b256-904b-4d8d-8c98-d829582c6739