Stop duplicating memory unnecessarily
Object names are immutable. When we're assigning strings from one
object (which has already been registered and added and can be found
in the object store), there's really no need for us to duplicate that
object's name in the slave object (such as hostnames for escalations
and dependencies, and timeperiod names for hosts and services and
escalations and whatnot), and we can avoid doing so most of the time,
actually reducing the code-count and making the remaining code simpler.

This should reduce memory usage and fragmentation quite a lot, and as
a nice bonus we no longer have to verify all objects so carefully
once they've been registered, and escalations and dependencies need
no further verification at all, saving a considerable amount of time
at startup.

Without actually measuring, the code certainly feels a lot snappier
now that we don't have to call out for more memory quite so often.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2107 5f96b256-904b-4d8d-8c98-d829582c6739