Use one-time allocation of object tables
It makes very little sense to allocate memory for one object at a time
as we create them in-core from configuration when we know early on
how many objects there will be. With this patch, we pre-allocate a
continuous block of memory, just large enough to hold all objects, and
then take our new objects from that store as we go along and create
them, making error handling in the 'add' objects even more trivial
than it used to be.

We keep the 'next' pointers for now though. They're all over the
place and I assume quite a few modules use them too.

While at it, we make sure the object lists are available from the
entire code and remove their explicit declarations in randomf files
as a general cleanup.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2115 5f96b256-904b-4d8d-8c98-d829582c6739