Remove incomplete failure prediction code
Failure prediction analysis is difficult and requires knowledge of
plugin values. One of the very strong points about Nagios is that
it doesn't have to care about plugin output (beyond storing and showing
it), so adding failure prediction on a global scope makes very little
sense. It would be doable by something that recognizes performance data
and can spot trends in such data. That's an excellent task for an
addon, but we'd do better not to riddle the core with such code.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2118 5f96b256-904b-4d8d-8c98-d829582c6739