xdata/xodtemplate: Fix group recombobulation
Previously we used to strcat() increasingly huge strings together.

With this patch, we create lists immediately instead. We could
actually improve even further upon the config parsing time by
registering group-contained objects before we recombobulate the
groups, but with this patch things seem pretty fast enough.

The one downside is that members are written in reverse order
in the precache and objects.cache files, but nothing (that I
know of) really cares about that, so I won't either. If it turns
out to be a problem, we can use the slow linked-list insertion
sort already present, but that would be slower.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2122 5f96b256-904b-4d8d-8c98-d829582c6739