xdata/xodtemplate: Fix host and service dependency counting
Previously we would allocate too little space for dependencies if
they were assigned to groups, since we failed to take into account
the duplicated dependencies. This patch fixes that, avoiding the
invalid writes we issued earlier that lead to (very) undefined
behaviour when we tried to handle whatever data we accidentally
overwrote.

Many thanks to Mark Frost for helping me locate this bug.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2124 5f96b256-904b-4d8d-8c98-d829582c6739