lib/bitmap: Allow unions between one set and the empty set
Previously we required two valid bitmaps to be passed to bitmap_union
and bitmap_unite, but taking the union of a set and the empty (NULL)
set is a valid operation that yields "the other" set. unioning two
empty sets ofcourse yields the empty set, so we're covered from all
angles.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2127 5f96b256-904b-4d8d-8c98-d829582c6739