Revert to old-school group recombobulation
Group recombobulation turned out to be a tad trickier than I first
thought, and while it's broken a lot of people are unable to test
Nagios 4 in any sort of long-running scenarios.

Unortunately, the dfs dependency patch ended up with the first
implementation on one side of the buggy recombobulation patch and
the fix of the first implementation on the other side, meaning
noone could really test any of it properly.

To get a clean revert, we also remove the buildfix for the group
recombobulation patch.

This reverts the following commits:
2059816f862b36598ce5301156f8a3c621712f9a
  Make cgi's compile with new group recombobulation code
abe461837d1a0955ef64d714b318e11a1a19cf18
  xdata/xodtemplate: Fix group recombobulation"

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2129 5f96b256-904b-4d8d-8c98-d829582c6739