base/config: Fix free()'ing of arrays in pre_flight_circular_check()
With high levels of optimization, the compiler would unroll the loop
and could yield code that tries to free() array subscripts with
negative numbers. We restrict the test to make sure the index never
goes below 0 and decrement it before checking to restrict the
compiler not to unroll that particular loop.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2130 5f96b256-904b-4d8d-8c98-d829582c6739