Reduce memory usage when parsing config
Using full integers to mark the "have's" of objects makes little sense,
since we only use one byte of them. With this patch, we reduce the
size of all xodtemplate_* objects. A lot of hosts and services and a
little for the others (done mostly by sed, so some of it makes very
little sense, but still get applied). We do this by adding bitfields
for the true booleans that never get passed by address (since bitfields
don't have addresses), saving us a neat 17megs of max rss for 10k
hosts and 100k services. Not too shabby for two minutes of work with
sed.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2135 5f96b256-904b-4d8d-8c98-d829582c6739