lib/squeue: Make sure we work well on 32-bit systems
Previously on 32-bit archs gcc would issue the following warning:
	squeue.c:34:  warning: left shift count >= width of type

which is legitimate, because unless we explicitly cast the
return type we'll do math on 32-bit values, which clearly can't
be shifted 32 bits or more.

With this patch we break the math apart so we always operate on a
type large enough to support it. For those weird systems where
'unsigned long long' is not 8 bytes wide, we do our best by simply
using the 32 bits at our disposal to store the timestamp.

Reported-by: Ricardo Maraschini <ricardo.maraschini@opservices.com.br>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2140 5f96b256-904b-4d8d-8c98-d829582c6739