lib/dkhash: Don't oversize tables so enormously
For small inputs, rounding up to the nearest power of two seemed like
a sensible idea, when it quickly escalates the memory usage beyond
reasonable limits when we have, say, 1.1million objects (when we would
allocate 2^21 buckets for the hash table, ending up overallocating by
a *huge* amount).

With this patch we pay the 'modulo price' and let the caller decide
the size of the hash table, which should (for preference) be x1.5
of the number of objects one wishes to store in it.

While at it, we refactor the dkhash_create() routine a bit to make
it less deeply indented.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2141 5f96b256-904b-4d8d-8c98-d829582c6739