core: Initialize service check result struct before calling modules
The earlier change (308a4221a62c67fdfd6aaa0b5bd6992896ca89dc, r1887)
had the intention of not setting the service as executing before we
called out to the module code, which was good, but we also changed
it to initialize the check result struct after the eventbroker
callbacks were made, which was bad.

This patch keeps the goodness and removes the badness, thereby
most likely fixing issue 350, though I haven't tested it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2148 5f96b256-904b-4d8d-8c98-d829582c6739