macros: Don't export member macros to the environment in large installs
Doing so will quickly eat up the available environment space, which is
shared by environment variables and program arguments, and lead to
return codes of 127 for checks and notifications regarding hosts and
services in a group with enough members with enough long name that a
comma-separated list of all member names consume too much of the
available (usually 128kbytes) memory reserved by the kernel for
program arguments and environment variables.

Users that desperately need that information will have to pass it
as command line arguments instead, and just keep their groups small
enough that their members and everything else can fit in the limited
memory the kernel gives us.

This provides an easy workaround for issue 354.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2149 5f96b256-904b-4d8d-8c98-d829582c6739