Make current state logging on log rotation configurable
It can cause an ungodly amount of spam in the logfiles, so we make
it configurable.

This patch was sent to the nagios tracker by one "awiddersheim",
but seems to have originally been coded up by one "ricardo". I'm
not sure, but I believe that's Ricardo Maraschini.

Issue #371 is hereby fixed.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2150 5f96b256-904b-4d8d-8c98-d829582c6739