core: Avoid scheduling stuff a week into the future
One pretty major issue with scheduling one week into the future is
that by next week the time will be exactly the same as it is now,
so if timeperiods are defined in a "normal" manner we'll be in the
exact same situation then as we are now, and we'll just keep on
scheduling stuff far into the future.

Another issue is that it's unlikely that checks are actually
supposed to run that far into the future, and we can therefore
assume that things have gone haywire beyond the usual manners.

The most sensible (for a given value of sensible) thing we can do
is to use the preferred time to schedule the check and see then
if we can get a sane timeslot for the check to run. That will
cause us to run into the event more frequently than necessary,
but it's not that great a chore and it will at least prevent
checks from going awol in scheduling limbo for any extended period
of time.

As a nice bonus, it also lets me close issue 31 for good and all,
as users now have a way to tweak the amount of time into the
future we'll schedule the checks when we fail to schedule them
due to some bugs, which is a Good Thing(tm) indeed and lets them
work around future problems by using some suitable number for
check interval. So long as it's shorter than the configured time
period slot, the check is guaranteed to run at least once.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2152 5f96b256-904b-4d8d-8c98-d829582c6739