core: Avoid bus errors and sigsegvs on weird configurations
When servicedependencies are created with dependent_hostgroup_name
and/or hostgroup_name and the groups mentioned there are empty, we
could previously bomb out with an illegal memory access when verifying
cyclic dependency paths.

With this patch, we provide belts, suspenders, helmet and condoms to
protect against such mishaps. Firstly, we set object tables created
with zero elements to NULL instead of just returning blindly, thereby
avoiding memory taints.

We also iterate using the object count and the static table rather
than list pointers, avoiding memory taints (but not counting errors),
and finally we make sure the dependency type is within the valid
range before we attempt to assign an array for checking the invalid
dependency.

Many, many thanks to Daniel Wittenberg who was invaluable in tracking
down and debugging this problem.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2155 5f96b256-904b-4d8d-8c98-d829582c6739